==> 1_extracting_IDs_from_zip.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Get-the-directories-within-the-zip-files-to-extract" data-toc-modified-id="Get-the-directories-within-the-zip-files-to-extract-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Get the directories within the zip files to extract</a></span></li><li><span><a href="#Extracting-sample-information-(aka-metadata)-and-BIOM-files" data-toc-modified-id="Extracting-sample-information-(aka-metadata)-and-BIOM-files-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Extracting sample information (aka metadata) and BIOM files</a></span></li></ul></div>

import pandas as pd
import os
import subprocess

# Get the directories within the zip files to extract

pwd

df = pd.read_csv('Literature_Sources-Sheet1.csv')
df = df[(~df['StudyID'].isna()) & (df['ID Final table'] != '-')]
df['StudyID'] = df['StudyID'].astype(int).astype(str)
list_dirs_to_extract = list(zip(df['StudyID'], df['ID Final table']))
print(len(list_dirs_to_extract))

# Extracting sample information (aka metadata) and BIOM files

for d in list_dirs_to_extract:
    print(d)
    os.chdir(d[0])
    cmd_1 = f"unzip sample_information*"
    subprocess.call(cmd_1, shell=True)
    cmd_2 = f'unzip biom_{d[0]}.zip "BIOM/{d[1]}/*"'
    subprocess.call(cmd_2, shell=True)
    # print(cmd_1)
    # print(cmd_2)
    os.chdir("../")
