==> 2_filter_control_samples_per_study.ipynb <==
I need to manually check each metadata because they entered controls in different ways...

Studies 10321 and 13450 don't seem to have a control.

import pandas as pd
pd.__version__

id714 = pd.read_csv('714/templates/714_20191015-105331.txt', sep = '\t')
id714 = id714[id714['treatment']=='control'].reset_index(drop=True)
id714.to_csv('714/templates/control_714.txt', sep='\t', index=False)
id714[['sample_name']].to_csv('714/templates/samples_to_keep_714.tsv', sep='\t', index=False)
id714.head()

id1609 = pd.read_csv('1609/templates/1609_20181018-131039.txt', sep='\t')
id1609 = id1609[id1609['control']=='y'].reset_index(drop=True)
id1609.to_csv('1609/templates/control_1609.txt', sep='\t', index=False)
id1609[['sample_name']].to_csv('1609/templates/samples_to_keep_1609.tsv', sep='\t', index=False)
id1609.head()

id10141 = pd.read_csv('10141/templates/10141_20190104-105500.txt', sep='\t')
id10141 = id10141[id10141['sample_type']=='control'].reset_index(drop=True)
id10141.to_csv('10141/templates/control_10141.txt', sep='\t', index=False)
id10141[['sample_name']].to_csv('10141/templates/samples_to_keep_10141.tsv', sep='\t', index=False)
id10141.head()

id10143 = pd.read_csv('10143/templates/10143_20191021-130723.txt', sep='\t')
id10143 = id10143[id10143['sample_name'].str.contains('ctrl')].reset_index(drop=True)
id10143.to_csv('10143/templates/control_10143.txt', sep='\t', index=False)
id10143[['sample_name']].to_csv('10143/templates/samples_to_keep_10143.tsv', sep='\t', index=False)
id10143.head()

## All samples comes from fly, and doesn't seem to have a control

# id10321 = pd.read_csv('10321/templates/10321_20180418-105405.txt', sep='\t')
# id10321
