==> 3_filter-samples_and_merge_biom.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#How-to-open-this-notebook?" data-toc-modified-id="How-to-open-this-notebook?-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>How to open this notebook?</a></span></li><li><span><a href="#714" data-toc-modified-id="714-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>714</a></span></li><li><span><a href="#1609" data-toc-modified-id="1609-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>1609</a></span></li><li><span><a href="#10141" data-toc-modified-id="10141-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>10141</a></span></li><li><span><a href="#10143" data-toc-modified-id="10143-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>10143</a></span></li><li><span><a href="#Merge-all" data-toc-modified-id="Merge-all-6"><span class="toc-item-num">6&nbsp;&nbsp;</span>Merge all</a></span></li><li><span><a href="#Extract-pmi_control-biom" data-toc-modified-id="Extract-pmi_control-biom-7"><span class="toc-item-num">7&nbsp;&nbsp;</span>Extract pmi_control biom</a></span></li><li><span><a href="#Merge-99_out_taxonomy.txt" data-toc-modified-id="Merge-99_out_taxonomy.txt-8"><span class="toc-item-num">8&nbsp;&nbsp;</span>Merge 99_out_taxonomy.txt</a></span></li></ul></div>

# How to open this notebook?

Need to use Qiime2, so need to open a notebook inside a conda environment:

```
[nathalia] $  conda activate qiime2-2020.11

(qiime2-2020.11) [nathalia] $  jnotebook 
```

Open this notebook - Bash kernel.

pwd

# 714

qiime tools import \
    --input-path 714/BIOM/46682/otu_table.biom \
    --type 'FeatureTable[Frequency]' \
    --input-format BIOMV210Format \
    --output-path 714/feature-table.qza

qiime feature-table filter-samples \
  --i-table 714/feature-table.qza \
  --m-metadata-file 714/templates/samples_to_keep_714.tsv \
  --o-filtered-table 714/feature-table_filtered.qza

# 1609

qiime tools import \
    --input-path 1609/BIOM/45905/otu_table.biom \
    --type 'FeatureTable[Frequency]' \
    --input-format BIOMV210Format \
    --output-path 1609/feature-table.qza

qiime feature-table filter-samples \
  --i-table 1609/feature-table.qza \
  --m-metadata-file 1609/templates/samples_to_keep_1609.tsv \
  --o-filtered-table 1609/feature-table_filtered.qza

# 10141

qiime tools import \
    --input-path 10141/BIOM/44878/otu_table.biom \
    --type 'FeatureTable[Frequency]' \
    --input-format BIOMV210Format \
    --output-path 10141/feature-table_44878.qza

qiime tools import \
    --input-path 10141/BIOM/44906/otu_table.biom \
    --type 'FeatureTable[Frequency]' \
    --input-format BIOMV210Format \
    --output-path 10141/feature-table_44906.qza

qiime feature-table merge \
    --i-tables 10141/feature-table_44878.qza \
    --i-tables 10141/feature-table_44906.qza \
    --p-overlap-method sum \
    --o-merged-table 10141/feature-table_merged.qza

qiime feature-table filter-samples \
  --i-table 10141/feature-table_merged.qza \
  --m-metadata-file 10141/templates/samples_to_keep_10141.tsv \
  --o-filtered-table 10141/feature-table_filtered.qza

# 10143

qiime tools import \
    --input-path 10143/BIOM/45990/otu_table.biom \
    --type 'FeatureTable[Frequency]' \
    --input-format BIOMV210Format \
    --output-path 10143/feature-table_45990.qza

qiime tools import \
    --input-path 10143/BIOM/46002/otu_table.biom \
    --type 'FeatureTable[Frequency]' \
    --input-format BIOMV210Format \
    --output-path 10143/feature-table_46002.qza

qiime feature-table merge \
    --i-tables 10143/feature-table_45990.qza \
    --i-tables 10143/feature-table_46002.qza \
    --p-overlap-method sum \
    --o-merged-table 10143/feature-table_merged.qza

qiime feature-table filter-samples \
  --i-table 10143/feature-table_merged.qza \
  --m-metadata-file 10143/templates/samples_to_keep_10143.tsv \
  --o-filtered-table 10143/feature-table_filtered.qza

# Merge all 

qiime feature-table merge \
    --i-tables 714/feature-table_filtered.qza \
    --i-tables 1609/feature-table_filtered.qza \
    --i-tables 10141/feature-table_filtered.qza \
    --i-tables 10143/feature-table_filtered.qza \
    --p-overlap-method sum \
    --o-merged-table pmi_control_feature-table.qza

# Extract pmi_control biom

unzip -K pmi_control_feature-table.qza

mv 95347b54-fc9f-4f83-b219-4cae5039cd7f pmi_control_feature-table_directory

biom convert -i pmi_control_feature-table_directory/data/feature-table.biom \
             -o pmi_control_feature-table.tsv --to-tsv
