==> 4_merge_with_otu_taxonomy.ipynb <==
import pandas as pd
import numpy as np

pmi = pd.read_csv('pmi_control_feature-table.tsv', sep="\t", skiprows=1)
pmi = pmi.replace(0, np.nan)
nsamples = pmi.shape[1]-1
psamples = int(nsamples*.8)
pmi = pmi.dropna(thresh=psamples).reset_index(drop=True)
pmi

taxonomy = pd.read_csv('99_otu_taxonomy.txt', sep='\t', header = None)
taxonomy = taxonomy.rename(columns={0:'#OTU ID', 1:'Taxonomy'})
taxonomy.head()

merged = pmi.merge(taxonomy, on='#OTU ID').replace(np.nan, 0)
merged.to_csv('pmi_taxonomy_control_samples.csv', index=False)
merged
